+++ src/Symfony/Component/DependencyInjection/Loader/Configurator/ContainerConfigurator.php
@@ -73,8 +73,6 @@ class ContainerConfigurator extends AbstractConfigurator
 
     final public function services(): ServicesConfigurator
     {
-        $this->loader->resetBeforeConfiguringServices();
-
         return new ServicesConfigurator($this->container, $this->loader, $this->instanceof, $this->path, $this->anonymousCount);
     }
 }