+++ src/Symfony/Component/DependencyInjection/Loader/Configurator/ServicesConfigurator.php
@@ -105,7 +105,6 @@ class ServicesConfigurator extends AbstractConfigurator
         $ref = static::processValue($referencedId, true);
         $alias = new Alias((string) $ref, $this->defaults->isPublic());
         $this->container->setAlias($id, $alias);
-        $this->loader->removeSinglyImplementedAlias((string) $ref);
 
         return new AliasConfigurator($this, $alias);
     }
@@ -139,5 +138,10 @@ class ServicesConfigurator extends AbstractConfigurator
     final public function __invoke(string $id, string $class = null): ServiceConfigurator
     {
         return $this->set($id, $class);
+    }
+
+    public function __destruct()
+    {
+        $this->loader->registerAliasesForSinglyImplementedInterfaces();
     }
 }