+++ src/Symfony/Component/DependencyInjection/Loader/YamlFileLoader.php
@@ -150,9 +150,7 @@ class YamlFileLoader extends FileLoader
             $this->parseDefinitions($content, $path);
         } finally {
             $this->instanceof = [];
-            $this->interfaces = [];
-            $this->singlyImplemented = [];
-            $this->singlyImplementedAliases = [];
+            $this->registerAliasesForSinglyImplementedInterfaces();
         }
     }
 
@@ -321,7 +319,6 @@ class YamlFileLoader extends FileLoader
 
         if (\is_string($service) && 0 === strpos($service, '@')) {
             $this->container->setAlias($id, $alias = new Alias(substr($service, 1)));
-            unset($this->singlyImplementedAliases[$id]);
             if (isset($defaults['public'])) {
                 $alias->setPublic($defaults['public']);
             }
@@ -345,7 +342,6 @@ class YamlFileLoader extends FileLoader
 
         if (isset($service['alias'])) {
             $this->container->setAlias($id, $alias = new Alias($service['alias']));
-            unset($this->singlyImplementedAliases[$id]);
             if (\array_key_exists('public', $service)) {
                 $alias->setPublic($service['public']);
             } elseif (isset($defaults['public'])) {